pyproject.toml
[tool.poetry]
name = "ratingsite"
version = "0.1.0"
description = ""
authors = ["Artemowkin <artyom.groshev2017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
PyYAML = "^6.0"
SQLAlchemy-Utils = "^0.38.2"
passlib = "^1.7.4"
aiopg = {extras = ["sa"], version = "^1.3.3"}
aiohttp_security = "^0.4.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
